Flytta noen funksjoner til RegistrationHelper. Nullstilling sletter nå registreringa
